fix: database persistence test with front and server changes

+++ myshoop/back_end/prisma/seed.ts
@@ -27,6 +27,10 @@ type picture = {
 }
 
 async function main() {
+  const store = await prisma.store.findFirst({})
+
+  if( store ) return;
+
   await prisma.publi.deleteMany({})
   await prisma.urlImage.deleteMany({})
   await prisma.cart.deleteMany({})
